requirements.txt
Flask==3.1.0
Werkzeug==3.1.2 
Flask-Cors==5.0.1
python-dotenv==1.0.1
instagrapi==2.1.3
gunicorn==20.1.0
